CatRecognition2/requirements.txt
# torch==2.5.1+cu121
# torchaudio==2.5.1+cu121
# torchvision==0.20.1+cu121
numpy==1.26.3
pillow==10.2.0
matplotlib==3.10.0
pandas==2.2.3
seaborn==0.13.2
scikit-learn==1.6.1
matplotlib
torch
torchvision
requests
tqdm
# torch==2.5.1+cu121 (download from website)